requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=a.txt ./requirements.txt
#
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
filelock==3.16.0
    # via torch
fsspec==2024.9.0
    # via torch
idna==3.10
    # via requests
jinja2==3.1.4
    # via torch
llvmlite==0.43.0
    # via numba
markupsafe==2.1.5
    # via jinja2
more-itertools==10.5.0
    # via openai-whisper
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numba==0.60.0
    # via openai-whisper
numpy==2.0.2
    # via
    #   numba
    #   openai-whisper
openai-whisper==20231117
    # via -r ./requirements.txt
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via tiktoken
sympy==1.13.2
    # via torch
tiktoken==0.7.0
    # via openai-whisper
torch==2.4.1
    # via openai-whisper
tqdm==4.66.5
    # via openai-whisper
typing-extensions==4.12.2
    # via torch
urllib3==2.2.3
    # via requests
